/feature_route_details_editor

+++ app/src/main/java/com/example/gh_coursework/data/TravelRepositoryImpl.kt
@@ -27,7 +27,7 @@ class TravelRepositoryImpl(private val localDataSrcIml: TravelDatasource.Local) 
         localDataSrcIml.addPointsTagsList(pointsTagsList)
     }
 
-    override suspend fun deletePoint(pointId: Int) {
+    override suspend fun deletePoint(pointId: Long) {
         localDataSrcIml.deletePoint(pointId)
     }
