+++ app/src/main/java/com/example/gh_coursework/data/datasource/TravelDatasource.kt
@@ -8,7 +8,7 @@ interface TravelDatasource {
     interface Local {
         suspend fun addOrUpdatePointOfInterestDetails(poi: PointDetailsDomain)
         suspend fun addPointOfInterestCoordinates(poi: PointPreviewDomain)
-        suspend fun deletePoint(pointId: Int)
+        suspend fun deletePoint(pointId: Long)
 
         suspend fun addPointTag(tag: PointTagDomain)
         suspend fun addPointsTagsList(pointsTagsList: List<PointsTagsDomain>)
