+++ app/src/main/java/com/example/gh_coursework/domain/entity/PointsTagsDomain.kt
@@ -2,5 +2,5 @@ package com.example.gh_coursework.domain.entity
 
 data class PointsTagsDomain(
     val pointId: Long,
-    val tagId: Int
+    val tagId: Long
 )
