+++ app/src/main/java/com/example/gh_coursework/domain/usecase/point_preview/DeletePointUseCaseImpl.kt
@@ -3,7 +3,7 @@ package com.example.gh_coursework.domain.usecase.point_preview
 import com.example.gh_coursework.domain.repository.TravelRepository
 
 class DeletePointUseCaseImpl(private val repository: TravelRepository) : DeletePointUseCase {
-    override suspend fun invoke(pointId: Int) {
+    override suspend fun invoke(pointId: Long) {
         repository.deletePoint(pointId)
     }
 }
