+++ app/src/main/java/com/example/gh_coursework/ui/point_details/model/PointsTagsModel.kt
@@ -2,5 +2,5 @@ package com.example.gh_coursework.ui.point_details.model
 
 data class PointsTagsModel(
     val pointId: Long,
-    val tagId: Int
+    val tagId: Long
 )
