+++ app/src/main/java/com/example/gh_coursework/ui/private_point/PrivatePointsFragment.kt
@@ -248,7 +248,7 @@ class PrivatePointsFragment : Fragment(R.layout.fragment_private_points) {
             }
 
             pointDetailsDeleteButton.setOnClickListener {
-                pointAnnotation.getData()?.asInt?.let { pointId ->
+                pointAnnotation.getData()?.asLong?.let { pointId ->
                     viewModel.deletePoint(
                         pointId
                     )
