+++ app/src/main/java/com/example/gh_coursework/ui/private_point/model/PrivatePointModel.kt
@@ -1,7 +1,7 @@
 package com.example.gh_coursework.ui.private_point.model
 
 data class PrivatePointModel(
-    val pointId: Int?,
+    val pointId: Long?,
     val x: Double,
     val y: Double,
     val isRoutePoint: Boolean
